refactor: clarify request and asset service responsibilities
Centralize membership request cancellation, snapshots, and deletion cleanup in request_service. Resolve invite recipients and join groups in routes before creating requests while preserving authorization and validation order.

Move asset ownership transfer validation into asset_service and reuse shared membership and request cleanup during group deletion. Add regression coverage for resolver snapshots and lookup ordering.
